Patch works great!
Product updates like a dream.

I used the inelegant getElementById method to extract data from the
modal to put into the patch, passing along a hidden store that holds
the id of the current object.

Send that to patch, good to go.

I know there is a better way of doing that, but for the sake of time
elegance is sacrificed for workhorse code.