Delete works
Delete will full delete the product. This raises concerns of what will
happen if an item is deleted and how that affects past orders.

The most viable solution that I can think of is to allow a new attribute
to the product modal called status and make it not required. If status
is hidden, however, then assign status to a class in tr to keep the
product and hide it from the cart.