Fixed the issue of delete
So when a full delete happens, it erases the product from the back end.
This will mess up the previous orders, so in order to over come that,
I put a drop menu inside the seller modal to allow the seller to choose
to instead hide the item from potential buyers. This will not affect
orders.

This is not the solution I would take in real life, as you may still
find a way through curl to access the item to add to an order, but
I would counter that through adding a limit as a seller.

Requiring this would in fact reformat the entire project, and that is
something I don not want to do right now as it is a group project and
that is not part of my mvp.