Fixed delete and disable
--The load checkout page button will now fully filter through store.cart
and check to see if any items in the list have a status different from
.hidden. This is becasue not all of our original data have a default
status. After this happens, it will go through and push any of those in
to a new array and then set that new array to store.cart and then fully
patch that before populating the checkout.

I am aware that this is sudo unnesscessary, but as we are not shutting
down our site for maintence the there is no worries about it. The best
practice would be to have a quantity counter and then wait some time
until after it has had it's run before disabling it.

-Deleting has a similiar affect as disabling and this fix fixes both.

Again this is bad practice as you never want to delete data that is
or has been used off your db without a backup record. For the sake of
our sites "story", I'm allowing this to work this way so any item that
has been deleted will now reflect as "Harmless kittens" because that is
awesome.